Refactor of TGN and Standardizer improvements
This commit is a first draft on improvements to the Geomash
gem. Very rough. Further improvements and clean up to come
this week. In essence:

1. TGN Sparql requests have been broken up. This is due to
TGN linked vocab site becomming too slow to do it all in one
Sparql query now. Overall it is now faster but requires more
HTTP hits.

2. Some slight improvements to standardizing of TGN terms.
Further improvements to come in that regard...